Clean up quiz validation component a bit
